Moved existing specs into the Cosmos namespace.
