+++ Solutions/Marain.Workflow.Specs/Features/Cosmos/Conditions.feature
