+++ Solutions/Marain.Workflow.Specs/Features/Cosmos/Delete.feature
