+++ Solutions/Marain.Workflow.Specs/Features/Cosmos/Delete.feature.cs
@@ -10,7 +10,7 @@
 // ------------------------------------------------------------------------------
 #region Designer generated code
 #pragma warning disable
-namespace Marain.Workflows.Specs.Features
+namespace Marain.Workflows.Specs.Features.Cosmos
 {
     using TechTalk.SpecFlow;
     using System;
