+++ Solutions/Marain.Workflow.Specs/Features/Cosmos/InvokeExternalServiceCondition.feature
