+++ Solutions/Marain.Workflow.Specs/Features/Cosmos/Publish.feature.cs
@@ -10,7 +10,7 @@
 // ------------------------------------------------------------------------------
 #region Designer generated code
 #pragma warning disable
-namespace Marain.Workflows.Specs.Features
+namespace Marain.Workflows.Specs.Features.Cosmos
 {
     using TechTalk.SpecFlow;
     using System;
@@ -20,11 +20,11 @@ namespace Marain.Workflows.Specs.Features
     [System.CodeDom.Compiler.GeneratedCodeAttribute("TechTalk.SpecFlow", "3.1.0.0")]
     [System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
     [NUnit.Framework.TestFixtureAttribute()]
-    [NUnit.Framework.DescriptionAttribute("Deprecate an item")]
+    [NUnit.Framework.DescriptionAttribute("Publish an item")]
     [NUnit.Framework.CategoryAttribute("perFeatureContainer")]
     [NUnit.Framework.CategoryAttribute("useCosmosStores")]
     [NUnit.Framework.CategoryAttribute("setupTenantedCosmosContainers")]
-    public partial class DeprecateAnItemFeature
+    public partial class PublishAnItemFeature
     {
         
         private TechTalk.SpecFlow.ITestRunner testRunner;
@@ -34,14 +34,14 @@ namespace Marain.Workflows.Specs.Features
                 "useCosmosStores",
                 "setupTenantedCosmosContainers"};
         
-#line 1 "Deprecated.feature"
+#line 1 "Publish.feature"
 #line hidden
         
         [NUnit.Framework.OneTimeSetUpAttribute()]
         public virtual void FeatureSetup()
         {
             testRunner = TechTalk.SpecFlow.TestRunnerManager.GetTestRunner();
-            TechTalk.SpecFlow.FeatureInfo featureInfo = new TechTalk.SpecFlow.FeatureInfo(new System.Globalization.CultureInfo("en-US"), "Deprecate an item", null, ProgrammingLanguage.CSharp, new string[] {
+            TechTalk.SpecFlow.FeatureInfo featureInfo = new TechTalk.SpecFlow.FeatureInfo(new System.Globalization.CultureInfo("en-US"), "Publish an item", null, ProgrammingLanguage.CSharp, new string[] {
                         "perFeatureContainer",
                         "useCosmosStores",
                         "setupTenantedCosmosContainers"});
@@ -83,13 +83,15 @@ namespace Marain.Workflows.Specs.Features
         }
         
         [NUnit.Framework.TestAttribute()]
-        [NUnit.Framework.DescriptionAttribute("Deprecate item when it is in the published state")]
+        [NUnit.Framework.DescriptionAttribute("Publish item when it is in the waiting for documentation state and it is complete" +
+            "")]
         [NUnit.Framework.CategoryAttribute("useChildObjects")]
-        public virtual void DeprecateItemWhenItIsInThePublishedState()
+        public virtual void PublishItemWhenItIsInTheWaitingForDocumentationStateAndItIsComplete()
         {
             string[] tagsOfScenario = new string[] {
                     "useChildObjects"};
-            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Deprecate item when it is in the published state", null, new string[] {
+            TechTalk.SpecFlow.ScenarioInfo scenarioInfo = new TechTalk.SpecFlow.ScenarioInfo("Publish item when it is in the waiting for documentation state and it is complete" +
+                    "", null, new string[] {
                         "useChildObjects"});
 #line 7
 this.ScenarioInitialize(scenarioInfo);
@@ -117,80 +119,72 @@ this.ScenarioInitialize(scenarioInfo);
 #line 9
  testRunner.And("the workflow trigger queue is ready to process new triggers", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "And ");
 #line hidden
-                TechTalk.SpecFlow.Table table20 = new TechTalk.SpecFlow.Table(new string[] {
+                TechTalk.SpecFlow.Table table49 = new TechTalk.SpecFlow.Table(new string[] {
                             "Key",
                             "Value"});
-                table20.AddRow(new string[] {
+                table49.AddRow(new string[] {
                             "Identifier",
                             "identifier1"});
-                table20.AddRow(new string[] {
+                table49.AddRow(new string[] {
                             "Type",
                             "t1"});
-                table20.AddRow(new string[] {
+                table49.AddRow(new string[] {
                             "AllowWaitingForDocumentationExit",
                             "x"});
-                table20.AddRow(new string[] {
+                table49.AddRow(new string[] {
                             "AllowPublishedEntry",
                             "x"});
 #line 10
  testRunner.And("I have created and persisted a new instance with Id \'id1\' of the workflow with Id" +
-                        " \'dc-workflow\' and supplied the following context items", ((string)(null)), table20, "And ");
+                        " \'dc-workflow\' and supplied the following context items", ((string)(null)), table49, "And ");
 #line hidden
-                TechTalk.SpecFlow.Table table21 = new TechTalk.SpecFlow.Table(new string[] {
+                TechTalk.SpecFlow.Table table50 = new TechTalk.SpecFlow.Table(new string[] {
                             "Id",
                             "Notes",
                             "Description"});
-                table21.AddRow(new string[] {
+                table50.AddRow(new string[] {
                             "id1",
                             "The new notes",
                             "The new description"});
 #line 16
  testRunner.And("I have an object of type \'application/vnd.endjin.datacatalog.catalogitempatchdeta" +
-                        "ils\' called \'patch\'", ((string)(null)), table21, "And ");
+                        "ils\' called \'patch\'", ((string)(null)), table50, "And ");
 #line hidden
-                TechTalk.SpecFlow.Table table22 = new TechTalk.SpecFlow.Table(new string[] {
+                TechTalk.SpecFlow.Table table51 = new TechTalk.SpecFlow.Table(new string[] {
                             "PatchDetails"});
-                table22.AddRow(new string[] {
+                table51.AddRow(new string[] {
                             "{patch}"});
 #line 19
  testRunner.And("I have sent the workflow engine a trigger of type \'application/vnd.endjin.datacat" +
-                        "alog.editcatalogitemtrigger\'", ((string)(null)), table22, "And ");
+                        "alog.editcatalogitemtrigger\'", ((string)(null)), table51, "And ");
 #line hidden
-                TechTalk.SpecFlow.Table table23 = new TechTalk.SpecFlow.Table(new string[] {
+                TechTalk.SpecFlow.Table table52 = new TechTalk.SpecFlow.Table(new string[] {
                             "CatalogItemId"});
-                table23.AddRow(new string[] {
+                table52.AddRow(new string[] {
                             "id1"});
 #line 22
- testRunner.And("I have sent the workflow engine a trigger of type \'application/vnd.endjin.datacat" +
-                        "alog.publishcatalogitemtrigger\'", ((string)(null)), table23, "And ");
-#line hidden
-                TechTalk.SpecFlow.Table table24 = new TechTalk.SpecFlow.Table(new string[] {
-                            "CatalogItemId"});
-                table24.AddRow(new string[] {
-                            "id1"});
-#line 25
  testRunner.When("I send the workflow engine a trigger of type \'application/vnd.endjin.datacatalog." +
-                        "deprecatecatalogitemtrigger\'", ((string)(null)), table24, "When ");
+                        "publishcatalogitemtrigger\'", ((string)(null)), table52, "When ");
 #line hidden
-#line 28
+#line 25
  testRunner.And("I wait for all triggers to be processed", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "And ");
 #line hidden
-#line 29
- testRunner.Then("the workflow instance with Id \'id1\' should have status \'Complete\'", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Then ");
+#line 26
+ testRunner.Then("the workflow instance with Id \'id1\' should have status \'Waiting\'", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "Then ");
 #line hidden
-#line 30
- testRunner.And("the workflow instance with Id \'id1\' should be in the state called \'Deprecated\'", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "And ");
+#line 27
+ testRunner.And("the workflow instance with Id \'id1\' should be in the state called \'Published\'", ((string)(null)), ((TechTalk.SpecFlow.Table)(null)), "And ");
 #line hidden
-                TechTalk.SpecFlow.Table table25 = new TechTalk.SpecFlow.Table(new string[] {
+                TechTalk.SpecFlow.Table table53 = new TechTalk.SpecFlow.Table(new string[] {
                             "Message"});
-                table25.AddRow(new string[] {
-                            "Exiting state \'Published\'"});
-                table25.AddRow(new string[] {
-                            "Executing transition \'Deprecate\'"});
-                table25.AddRow(new string[] {
-                            "Entering state \'Deprecated\'"});
-#line 31
- testRunner.And("the following trace messages should be the last messages recorded", ((string)(null)), table25, "And ");
+                table53.AddRow(new string[] {
+                            "Exiting state \'Waiting for documentation\'"});
+                table53.AddRow(new string[] {
+                            "Executing transition \'Publish\'"});
+                table53.AddRow(new string[] {
+                            "Entering state \'Published\'"});
+#line 28
+ testRunner.And("the following trace messages should be the last messages recorded", ((string)(null)), table53, "And ");
 #line hidden
             }
             this.ScenarioCleanup();
