+++ Solutions/Marain.Workflow.Specs/Marain.Workflow.Specs.csproj
@@ -56,7 +56,4 @@
     </None>
   </ItemGroup>
 
-  <ItemGroup>
-  </ItemGroup>
-
 </Project>
